Split func mod in replay and result
