+++ crates/evm/evm/src/executors/invariant/error.rs
@@ -1,18 +1,10 @@
-use super::{replay_run, BasicTxDetails, InvariantContract};
-use crate::executors::{invariant::shrink::shrink_sequence, Executor, RawCallResult};
-use alloy_primitives::{Address, Bytes, Log};
-use eyre::Result;
-use foundry_common::contracts::{ContractsByAddress, ContractsByArtifact};
+use super::{BasicTxDetails, InvariantContract};
+use crate::executors::RawCallResult;
+use alloy_primitives::{Address, Bytes};
 use foundry_config::InvariantConfig;
 use foundry_evm_core::decode::RevertDecoder;
-use foundry_evm_coverage::HitMaps;
-use foundry_evm_fuzz::{
-    invariant::FuzzRunIdentifiedContracts, CounterExample, FuzzedCases, Reason,
-};
-use foundry_evm_traces::{CallTraceArena, Traces};
-use parking_lot::RwLock;
+use foundry_evm_fuzz::{invariant::FuzzRunIdentifiedContracts, Reason};
 use proptest::test_runner::TestError;
-use std::sync::Arc;
 
 /// Stores information about failures and reverts of the invariant tests.
 #[derive(Clone, Default)]
@@ -35,23 +27,6 @@ impl InvariantFailures {
     pub fn into_inner(self) -> (usize, Option<InvariantFuzzError>) {
         (self.reverts, self.error)
     }
-}
-
-/// The outcome of an invariant fuzz test
-#[derive(Debug)]
-pub struct InvariantFuzzTestResult {
-    pub error: Option<InvariantFuzzError>,
-    /// Every successful fuzz test case
-    pub cases: Vec<FuzzedCases>,
-    /// Number of reverted fuzz calls
-    pub reverts: usize,
-
-    /// The entire inputs of the last run of the invariant campaign, used for
-    /// replaying the run for collecting traces.
-    pub last_run_inputs: Vec<BasicTxDetails>,
-
-    /// Additional traces used for gas report construction.
-    pub gas_report_traces: Vec<Vec<CallTraceArena>>,
 }
 
 #[derive(Clone, Debug)]
@@ -131,56 +106,6 @@ impl FailedInvariantCaseData {
             shrink_sequence: invariant_config.shrink_sequence,
             shrink_run_limit: invariant_config.shrink_run_limit,
             fail_on_revert: invariant_config.fail_on_revert,
-        }
-    }
-
-    /// Replays the error case, shrinks the failing sequence and collects all necessary traces.
-    #[allow(clippy::too_many_arguments)]
-    pub fn replay_error(
-        &self,
-        invariant_contract: &InvariantContract<'_>,
-        mut executor: Executor,
-        known_contracts: &ContractsByArtifact,
-        ided_contracts: ContractsByAddress,
-        logs: &mut Vec<Log>,
-        traces: &mut Traces,
-        coverage: &mut Option<HitMaps>,
-    ) -> Result<Option<CounterExample>> {
-        match self.test_error {
-            // Don't use at the moment.
-            TestError::Abort(_) => Ok(None),
-            TestError::Fail(_, ref calls) => {
-                // Shrink sequence of failed calls.
-                let calls = if self.shrink_sequence {
-                    shrink_sequence(self, calls, &executor)?
-                } else {
-                    trace!(target: "forge::test", "Shrinking disabled.");
-                    calls.clone()
-                };
-
-                set_up_inner_replay(&mut executor, &self.inner_sequence);
-                // Replay calls to get the counterexample and to collect logs, traces and coverage.
-                replay_run(
-                    invariant_contract,
-                    executor,
-                    known_contracts,
-                    ided_contracts,
-                    logs,
-                    traces,
-                    coverage,
-                    calls,
-                )
-            }
-        }
-    }
-}
-
-/// Sets up the calls generated by the internal fuzzer, if they exist.
-fn set_up_inner_replay(executor: &mut Executor, inner_sequence: &[Option<BasicTxDetails>]) {
-    if let Some(fuzzer) = &mut executor.inspector.fuzzer {
-        if let Some(call_generator) = &mut fuzzer.call_generator {
-            call_generator.last_sequence = Arc::new(RwLock::new(inner_sequence.to_owned()));
-            call_generator.set_replay(true);
         }
     }
 }
