+++ crates/evm/evm/src/executors/invariant/replay.rs
@@ -1,71 +1,20 @@
-use super::{error::FailedInvariantCaseData, InvariantFailures, InvariantFuzzError};
-use crate::executors::{Executor, RawCallResult};
+use super::{error::FailedInvariantCaseData, shrink_sequence};
+use crate::executors::Executor;
 use alloy_dyn_abi::JsonAbiExt;
 use alloy_primitives::Log;
 use eyre::Result;
 use foundry_common::{ContractsByAddress, ContractsByArtifact};
-use foundry_config::InvariantConfig;
 use foundry_evm_core::constants::CALLER;
 use foundry_evm_coverage::HitMaps;
 use foundry_evm_fuzz::{
-    invariant::{BasicTxDetails, FuzzRunIdentifiedContracts, InvariantContract},
+    invariant::{BasicTxDetails, InvariantContract},
     BaseCounterExample, CounterExample,
 };
 use foundry_evm_traces::{load_contracts, TraceKind, Traces};
+use parking_lot::RwLock;
+use proptest::test_runner::TestError;
 use revm::primitives::U256;
-use std::borrow::Cow;
-
-/// Given the executor state, asserts that no invariant has been broken. Otherwise, it fills the
-/// external `invariant_failures.failed_invariant` map and returns a generic error.
-/// Either returns the call result if successful, or nothing if there was an error.
-pub fn assert_invariants(
-    invariant_contract: &InvariantContract<'_>,
-    invariant_config: &InvariantConfig,
-    targeted_contracts: &FuzzRunIdentifiedContracts,
-    executor: &Executor,
-    calldata: &[BasicTxDetails],
-    invariant_failures: &mut InvariantFailures,
-) -> Result<Option<RawCallResult>> {
-    let mut inner_sequence = vec![];
-
-    if let Some(fuzzer) = &executor.inspector.fuzzer {
-        if let Some(call_generator) = &fuzzer.call_generator {
-            inner_sequence.extend(call_generator.last_sequence.read().iter().cloned());
-        }
-    }
-
-    let func = invariant_contract.invariant_function;
-    let mut call_result = executor.call_raw(
-        CALLER,
-        invariant_contract.address,
-        func.abi_encode_input(&[]).expect("invariant should have no inputs").into(),
-        U256::ZERO,
-    )?;
-
-    let is_err = !executor.is_raw_call_success(
-        invariant_contract.address,
-        Cow::Owned(call_result.state_changeset.take().unwrap()),
-        &call_result,
-        false,
-    );
-    if is_err {
-        // We only care about invariants which we haven't broken yet.
-        if invariant_failures.error.is_none() {
-            let case_data = FailedInvariantCaseData::new(
-                invariant_contract,
-                invariant_config,
-                targeted_contracts,
-                calldata,
-                call_result,
-                &inner_sequence,
-            );
-            invariant_failures.error = Some(InvariantFuzzError::BrokenInvariant(case_data));
-            return Ok(None);
-        }
-    }
-
-    Ok(Some(call_result))
-}
+use std::sync::Arc;
 
 /// Replays a call sequence for collecting logs and traces.
 /// Returns counterexample to be used when the call sequence is a failed scenario.
@@ -132,4 +81,54 @@ pub fn replay_run(
 
     Ok((!counterexample_sequence.is_empty())
         .then_some(CounterExample::Sequence(counterexample_sequence)))
+}
+
+/// Replays the error case, shrinks the failing sequence and collects all necessary traces.
+#[allow(clippy::too_many_arguments)]
+pub fn replay_error(
+    failed_case: &FailedInvariantCaseData,
+    invariant_contract: &InvariantContract<'_>,
+    mut executor: Executor,
+    known_contracts: &ContractsByArtifact,
+    ided_contracts: ContractsByAddress,
+    logs: &mut Vec<Log>,
+    traces: &mut Traces,
+    coverage: &mut Option<HitMaps>,
+) -> Result<Option<CounterExample>> {
+    match failed_case.test_error {
+        // Don't use at the moment.
+        TestError::Abort(_) => Ok(None),
+        TestError::Fail(_, ref calls) => {
+            // Shrink sequence of failed calls.
+            let calls = if failed_case.shrink_sequence {
+                shrink_sequence(failed_case, calls, &executor)?
+            } else {
+                trace!(target: "forge::test", "Shrinking disabled.");
+                calls.clone()
+            };
+
+            set_up_inner_replay(&mut executor, &failed_case.inner_sequence);
+            // Replay calls to get the counterexample and to collect logs, traces and coverage.
+            replay_run(
+                invariant_contract,
+                executor,
+                known_contracts,
+                ided_contracts,
+                logs,
+                traces,
+                coverage,
+                calls,
+            )
+        }
+    }
+}
+
+/// Sets up the calls generated by the internal fuzzer, if they exist.
+fn set_up_inner_replay(executor: &mut Executor, inner_sequence: &[Option<BasicTxDetails>]) {
+    if let Some(fuzzer) = &mut executor.inspector.fuzzer {
+        if let Some(call_generator) = &mut fuzzer.call_generator {
+            call_generator.last_sequence = Arc::new(RwLock::new(inner_sequence.to_owned()));
+            call_generator.set_replay(true);
+        }
+    }
 }
